monitor/requirements.txt
wmi
psutil
pillow
aiohttp